+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 > This project follows the [Keep a Changelog](https://keepachangelog.com/en/1.1.0/) format.
 
 ## [Unreleased]
+* rename `globalCache.clear()` to `globalCache.clearTestRun()`.
 
 ## [0.2.16] - 2025-10-15
 * fix: handle Playwright UI mode and VSCode extension runs.